+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#30](https://github.com/saas-rs/cli/issues/30) Support listing all features available for generation
 
 ### Fixed
+- [#32](https://github.com/saas-rs/cli/issues/32) The archive uploaded in a generate request should not be buffered in memory
 - [#22](https://github.com/saas-rs/cli/issues/22) version subcommand panics when installed via cargo
 
 ## [0.1.8] - 2025-06-06